Fixes for price type

+++ src/CrossContext/Hotelio.CrossContext.Contract/Pricing/IPricingService.cs
@@ -2,5 +2,5 @@ namespace Hotelio.CrossContext.Contract.Pricing;
 
 public interface IPricingService
 {
-    public Task<double> CalculatePrice(string hotelId, string roomType, List<string> amenities, int guests, DateTime startDate, DateTime endDate);
+    public Task<decimal> CalculatePrice(string hotelId, string roomType, List<string> amenities, int guests, DateTime startDate, DateTime endDate);
 }
