+++ src/Modules/Booking/Hotelio.Modules.Booking.Application/ReadModel/Reservation.cs
@@ -14,15 +14,15 @@ internal class Reservation
     public RoomType RoomType { get; set; }
     public int NumberOfGuests { get; set; }
     public string Status { get; set; }
-    public double PriceToPay { get; set; }
-    public double PricePayed { get; set; }
+    public decimal PriceToPay { get; set; }
+    public decimal PricePayed { get; set; }
     public string PaymentType { get; set; }
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
     public List<Amenity> Amenities { get; set; }
     public string? RoomId { get; set; }
 
-    public Reservation(string id, Hotel hotel, Owner owner, RoomType roomType, int numberOfGuests, string status, double priceToPay, double pricePayed, string paymentType, DateTime startDate, DateTime endDate, List<Amenity> amenities, string? roomId = null)
+    public Reservation(string id, Hotel hotel, Owner owner, RoomType roomType, int numberOfGuests, string status, decimal priceToPay, decimal pricePayed, string paymentType, DateTime startDate, DateTime endDate, List<Amenity> amenities, string? roomId = null)
     {
         Id = id;
         Hotel = hotel;
