+++ src/Modules/Pricing/Hotelio.Modules.Pricing.Application/Command/AddAmenityTariff.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace Hotelio.Modules.Pricing.Application.Command;
 
-internal record AddAmenityTariff(Guid HotelTariffId, string AmenityId, double PriceNetAmount) : IRequest;
+internal record AddAmenityTariff(Guid HotelTariffId, string AmenityId, decimal PriceNetAmount) : IRequest;
