+++ src/Modules/Pricing/Hotelio.Modules.Pricing.Application/Command/AddRoomPeriodPrice.cs
@@ -5,6 +5,6 @@ namespace Hotelio.Modules.Pricing.Application.Command;
 internal record AddRoomPeriodPrice(
     Guid HotelTariffId,
     string RoomTypeId,
-    double PriceNetAmount,
+    decimal PriceNetAmount,
     DateTime StartDate,
     DateTime EndDate) : IRequest;
