+++ src/Modules/Pricing/Hotelio.Modules.Pricing.Application/Command/CreateHotelTariff.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace Hotelio.Modules.Pricing.Application.Command;
 
-internal record CreateHotelTariff(string HotelId, double BasePriceNetAmount) : IRequest<Guid>;
+internal record CreateHotelTariff(string HotelId, decimal BasePriceNetAmount) : IRequest<Guid>;
